+++ packages/zowe-explorer/__tests__/__unit__/shared/init.unit.test.ts
@@ -26,6 +26,7 @@ import { saveUSSFile } from "../../../src/uss/actions";
 import { ZoweLogger } from "../../../src/utils/LoggerUtils";
 import { ZoweSaveQueue } from "../../../src/abstract/ZoweSaveQueue";
 import { ZoweExplorerApiRegister } from "../../../src/ZoweExplorerApiRegister";
+import * as HistoryView from "../../../src/shared/HistoryView";
 import { LocalFileManagement } from "../../../src/utils/LocalFileManagement";
 
 jest.mock("../../../src/utils/LoggerUtils");
@@ -55,11 +56,8 @@ describe("Test src/shared/extension", () => {
                 mock: [],
             },
             {
-                name: "zowe.updateSecureCredentials",
-                mock: [
-                    { spy: jest.spyOn(globals, "setGlobalSecurityValue"), arg: [test.value] },
-                    { spy: jest.spyOn(profUtils.ProfilesUtils, "writeOverridesFile"), arg: [] },
-                ],
+                name: "zowe.editHistory",
+                mock: [{ spy: jest.spyOn(HistoryView, "HistoryView"), arg: [test.context, test.value.providers] }],
             },
             {
                 name: "zowe.promptCredentials",
@@ -387,6 +385,24 @@ describe("Test src/shared/extension", () => {
             expect(context.subscriptions).toContain(treeView);
             expect(spyCollapse).not.toHaveBeenCalled();
             expect(spyExpand).not.toHaveBeenCalled();
+        });
+    });
+
+    describe("registerCredentialManager", () => {
+        let context: any;
+
+        beforeEach(() => {
+            context = { subscriptions: [] };
+            jest.clearAllMocks();
+        });
+        afterAll(() => {
+            jest.restoreAllMocks();
+        });
+
+        it("should register command for updating credentials", () => {
+            const registerCommandSpy = jest.spyOn(vscode.commands, "registerCommand");
+            sharedExtension.registerCredentialManager(context);
+            expect(registerCommandSpy).toBeCalledWith("zowe.updateSecureCredentials", expect.any(Function));
         });
     });
 });