+++ packages/zowe-explorer/src/abstract/ZoweTreeProvider.ts
@@ -86,6 +86,16 @@ export class ZoweTreeProvider {
     }
 
     /**
+     * Fire the "onDidChangeTreeData" event to signal that a node in the tree has changed.
+     * Unlike `refreshElement`, this function does *not* signal a refresh for the given node -
+     * it simply tells VS Code to repaint the node in the tree.
+     * @param node The node that should be repainted
+     */
+    public nodeDataChanged(node: IZoweTreeNode): void {
+        this.mOnDidChangeTreeData.fire(node);
+    }
+
+    /**
      * Called whenever the tree needs to be refreshed, and fires the data change event
      *
      */