+++ packages/zowe-explorer/src/shared/init.ts
@@ -27,6 +27,8 @@ import { ZoweLogger } from "../utils/LoggerUtils";
 import { ZoweSaveQueue } from "../abstract/ZoweSaveQueue";
 import { SettingsConfig } from "../utils/SettingsConfig";
 import { spoolFilePollEvent } from "../job/actions";
+import { HistoryView } from "./HistoryView";
+import { ProfileManagement } from "../utils/ProfileManagement";
 import { LocalFileManagement } from "../utils/LocalFileManagement";
 
 // Set up localization
@@ -79,17 +81,22 @@ export function registerCommonCommands(context: vscode.ExtensionContext, provide
         })
     );
 
-    // Update imperative.json to false only when VS Code setting is set to false
+    // Webview for editing persistent items on Zowe Explorer
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.updateSecureCredentials", async (customCredentialManager?: string) => {
-            await globals.setGlobalSecurityValue(customCredentialManager);
-            ProfilesUtils.writeOverridesFile();
+        vscode.commands.registerCommand("zowe.editHistory", () => {
+            return new HistoryView(context, providers);
         })
     );
 
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.promptCredentials", async (node: IZoweTreeNode) => {
             await ProfilesUtils.promptCredentials(node);
+        })
+    );
+
+    context.subscriptions.push(
+        vscode.commands.registerCommand("zowe.profileManagement", async (node: IZoweTreeNode) => {
+            await ProfileManagement.manageProfile(node);
         })
     );
 
@@ -224,6 +231,16 @@ export function registerCommonCommands(context: vscode.ExtensionContext, provide
         // initialize the globals.filesToCompare array during initialization
         globals.resetCompareChoices();
     }
+}
+
+export function registerCredentialManager(context: vscode.ExtensionContext): void {
+    // Update imperative.json to false only when VS Code setting is set to false
+    context.subscriptions.push(
+        vscode.commands.registerCommand("zowe.updateSecureCredentials", async (customCredentialManager?: string) => {
+            await globals.setGlobalSecurityValue(customCredentialManager);
+            ProfilesUtils.writeOverridesFile();
+        })
+    );
 }
 
 export function watchConfigProfile(context: vscode.ExtensionContext, providers: IZoweProviders): void {